requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.4
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-monthfield==0.1.4
django-static-jquery3==5.1.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
